hw4/requirements.txt
torch
torchvision
matplotlib
pillow
scikit-image
einops
piq